demo separate modals per report type
