+++ services/ui-src/src/components/index.ts
@@ -79,8 +79,8 @@ export { Modal } from "./modals/Modal";
 export { AddEditEntityModal } from "./modals/AddEditEntityModal";
 export { AddEditOverlayEntityModal } from "./modals/AddEditOverlayEntityModal";
 export { CloseEntityModal } from "./modals/CloseEntityModal";
-export { AddNewWorkPlanModal } from "./modals/AddNewWorkPlanModal";
-export { AddEditReportModal } from "./modals/AddEditReportModal";
+export { CreateSarModal } from "./modals/CreateSarModal";
+export { CreateWorkPlanModal } from "./modals/CreateWorkPlanModal";
 export { DeleteEntityModal } from "./modals/DeleteEntityModal";
 // pages
 export { AdminPage } from "./pages/Admin/AdminPage";
